Refactored code, created updateLabelsForChildControllers() func to update labels programmatically on InfoViewController
