+++ Contained/Contained/Base.lproj/Main.storyboard
@@ -32,28 +32,14 @@
             </objects>
             <point key="canvasLocation" x="864" y="567"/>
         </scene>
-        <!--View Controller-->
+        <!--Info View Controller-->
         <scene sceneID="r5h-ul-XOI">
             <objects>
-                <viewController id="acA-ae-cBD" sceneMemberID="viewController">
+                <viewController id="acA-ae-cBD" customClass="InfoViewController" customModule="Contained" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="zzb-kJ-sOa">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="0mf-ar-3r1">
-                                <rect key="frame" x="140" y="419.5" width="134.5" height="57.5"/>
-                                <string key="text">Tap to move
-the crab</string>
-                                <fontDescription key="fontDescription" type="boldSystem" pointSize="24"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                        </subviews>
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="0mf-ar-3r1" firstAttribute="centerX" secondItem="zzb-kJ-sOa" secondAttribute="centerX" id="I4X-5R-3x1"/>
-                            <constraint firstItem="0mf-ar-3r1" firstAttribute="centerY" secondItem="zzb-kJ-sOa" secondAttribute="centerY" id="atl-oe-DKu"/>
-                        </constraints>
                         <viewLayoutGuide key="safeArea" id="L0w-Lj-izN"/>
                     </view>
                     <navigationItem key="navigationItem" id="FUY-yo-5f5">
@@ -68,38 +54,25 @@ the crab</string>
             </objects>
             <point key="canvasLocation" x="1774" y="567"/>
         </scene>
-        <!--View Controller-->
+        <!--Info View Controller-->
         <scene sceneID="era-NL-wgC">
             <objects>
-                <viewController id="Ztx-V0-Vma" sceneMemberID="viewController">
+                <viewController id="Ztx-V0-Vma" customClass="InfoViewController" customModule="Contained" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="wsz-2b-t8W">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Let's play!" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="2" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="NC0-ad-yWB">
-                                <rect key="frame" x="149.5" y="433.5" width="115" height="29"/>
-                                <fontDescription key="fontDescription" type="boldSystem" pointSize="24"/>
-                                <nil key="textColor"/>
-                                <nil key="highlightedColor"/>
-                            </label>
-                        </subviews>
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstItem="NC0-ad-yWB" firstAttribute="centerY" secondItem="wsz-2b-t8W" secondAttribute="centerY" id="QQU-Dr-jAp"/>
-                            <constraint firstItem="NC0-ad-yWB" firstAttribute="centerX" secondItem="wsz-2b-t8W" secondAttribute="centerX" id="fjQ-oH-xTk"/>
-                        </constraints>
                         <viewLayoutGuide key="safeArea" id="ZmT-JV-Kdh"/>
                     </view>
                     <navigationItem key="navigationItem" id="xt4-nk-irZ">
                         <barButtonItem key="rightBarButtonItem" title="Done" id="kW5-d4-V0i">
                             <connections>
-                                <action selector="done:" destination="M4R-mz-jZb" id="NaM-TQ-hI2"/>
+                                <action selector="done:" destination="Ztx-V0-Vma" id="qZ4-oB-e9V"/>
                             </connections>
                         </barButtonItem>
                     </navigationItem>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="Ve0-P8-FMF" userLabel="First Responder" sceneMemberID="firstResponder"/>
-                <exit id="M4R-mz-jZb" userLabel="Exit" sceneMemberID="exit"/>
             </objects>
             <point key="canvasLocation" x="2713" y="567"/>
         </scene>
